% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_any_nslp.R
\name{nemwebR_any_nlsp}
\alias{nemwebR_any_nlsp}
\title{Retrieve Net System Load Profile data}
\usage{
nemwebR_any_nlsp(year)
}
\arguments{
\item{year}{integer of the form YYYY}
}
\value{
A data frame
}
\description{
This function returns Net System Load Profile (NSLP) data from AEMO's NEMWeb as specified by the year argument. If the current year is queried the function will retrieve all weeks in the year.
}
\details{
Net System Load Profiles are published for each distributor.

Data is available from 2002 to 2021. This code will need updating in 2022.
}
\examples{
nemwebR_any_nslp(2019)

}
