% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_biddayoffer.R
\name{nemwebR_archive_biddayoffer}
\alias{nemwebR_archive_biddayoffer}
\title{Retrieve archived BIDDAYOFFER data}
\usage{
nemwebR_archive_biddayoffer(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of BIDDAYOFFER data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
BIDDAYOFFER data contains bid price band data for all scheduled and non-scheduled generators in the NEM.

Archive data is available from July 2009 to approximately one month ago. In order to retrieve newer data you will need to use the nemwebR_current_BIDDAYOFFER function.
}
\examples{
nemwebR_archive_biddayoffer(20210101)

}
