% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_bidduiddetails.R
\name{nemwebR_archive_bidduiddetails}
\alias{nemwebR_archive_bidduiddetails}
\title{Retrieve archived BIDDUIDDETAILS data}
\usage{
nemwebR_archive_bidduiddetails(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMM}
}
\value{
A data frame
}
\description{
This function returns one month of BIDDUIDDETAILS data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
BIDDUIDDETAILS data contains relevant minimum and maximum enablement quantities and FCAS trapezium angles for all scheduled and non-scheduled generators in the NEM.

Archive data is available from July 2009 to approximately two weeks ago. In order to retrieve newer data you will need to use the nemwebR_current_BIDDUIDDETAILS function.
}
\examples{
nemwebR_archive_bidduiddetails(202101)

}
