% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_bidperoffer.R
\name{nemwebR_archive_bidperoffer}
\alias{nemwebR_archive_bidperoffer}
\title{Retrieve archived BIDPEROFFER data}
\usage{
nemwebR_archive_bidperoffer(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of BIDPEROFFER data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
BIDPEROFFER data contains bid energy and FCAS volumes for all scheduled and non-scheduled generators in the NEM.

Archive data is available from July 2009 to approximately one month ago. In order to retrieve newer data you will need to use the nemwebR_current_BIDPEROFFER function.
}
\examples{
nemwebR_archive_bidperoffer(20210101)

}
