% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_dispatch_fcas_req.R
\name{nemwebR_archive_dispatch_fcas_req}
\alias{nemwebR_archive_dispatch_fcas_req}
\title{Retrieve archived DISPATCH_FCAS_REQ data}
\usage{
nemwebR_archive_dispatch_fcas_req(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of DISPATCH_FCAS_REQ data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
DISPATCH_FCAS_REQ data contains historical 5-minute generation quantities for all scheduled and non-scheduled generators in the NEM.

Archive data is available from July 2009 to approximately one month ago. In order to retrieve newer data you will need to use the nemwebR_current_DISPATCH_FCAS_REQ function.
}
\examples{
nemwebR_archive_dispatch_fcas_req(20210101)

}
