% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_dispatchload.R
\name{nemwebR_archive_dispatchload}
\alias{nemwebR_archive_dispatchload}
\title{Retrieve archived DISPATCHLOAD data}
\usage{
nemwebR_archive_dispatchload(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of DISPATCHLOAD data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
DISPATCHLOAD data contains historical 5-minute generation quantities for all scheduled and non-scheduled generators in the NEM.

Archive data is available from July 2009 to approximately one month ago. In order to retrieve newer data you will need to use the nemwebR_current_DISPATCHLOAD function.
}
\examples{
nemwebR_archive_dispatchload(20210101)

}
