% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_dispatchprice.R
\name{nemwebR_archive_dispatchprice}
\alias{nemwebR_archive_dispatchprice}
\title{Retrieve archived DISPATCHPRICE data}
\usage{
nemwebR_archive_dispatchprice(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMM}
}
\value{
A data frame
}
\description{
This function returns one month of DISPATCHPRICE data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
DISPATCHPRICE data contains historical 5-minute generation quantities for all scheduled and non-scheduled generators in the NEM.

Archive data is available from July 2009 to approximately two weeks ago. In order to retrieve newer data you will need to use the nemwebR_current_DISPATCHPRICE function.
}
\examples{
nemwebR_archive_dispatchprice(202101)

}
