% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_dudetail.R
\name{nemwebR_archive_dudetail}
\alias{nemwebR_archive_dudetail}
\title{Retrieve archived DUDETAIL data}
\usage{
nemwebR_archive_dudetail(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of DUDETAIL data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
DUDETAIL data contains historical DUID registration details for all registered generators in the NEM.

Archive data is available from July 2009 to approximately one month ago. In order to retrieve newer data you will need to use the nemwebR_current_DUDETAIL function.
}
\examples{
nemwebR_archive_dispatch_dudetail(20210101)

}
