% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_dudetailsummary.R
\name{nemwebR_archive_dudetailsummary}
\alias{nemwebR_archive_dudetailsummary}
\title{Retrieve archived DUDETAILSUMMARY data}
\usage{
nemwebR_archive_dudetailsummary(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMM}
}
\value{
A data frame
}
\description{
This function returns one month of DUDETAILSUMMARY data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
DUDETAILSUMMARY data contains historical DUID registration details for all registered generators in the NEM.

Archive data is available from July 2009 to approximately two weeks ago. In order to retrieve newer data you will need to use the nemwebR_current_DUDETAILSUMMARY function.
}
\examples{
nemwebR_archive_dudetailsummary(202101)

}
