% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_genunits.R
\name{nemwebR_archive_genunits}
\alias{nemwebR_archive_genunits}
\title{Retrieve archived GENUNITS data}
\usage{
nemwebR_archive_genunits(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of GENUNITS data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
GENUNITS data contains historical generation unit registration details, including fuel type and CO2 emissions information.

Archive data is available from July 2009 to approximately one month ago. In order to retrieve newer data you will need to use the nemwebR_current_GENUNITS function.
}
\examples{
nemwebR_archive_dispatch_genunits(20210101)

}
