% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_tradinginterconnect.R
\name{nemwebR_archive_tradinginterconnect}
\alias{nemwebR_archive_tradinginterconnect}
\title{Retrieve archived TRADINGINTERCONNECT data}
\usage{
nemwebR_archive_tradinginterconnect(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of TRADINGINTERCONNECT data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
TRADINGINTERCONNECT data contains historical 30-minute interconnector limits and MW flows.

Archive data is available from July 2009 up to approximately one month ago. In order to retrieve newer data you will need to use the nemwebR_current_TRADINGINTERCONNECT function.
}
\examples{
nemwebR_archive_tradinginterconnect(20210101)

}
