% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_tradingload.R
\name{nemwebR_archive_tradingload}
\alias{nemwebR_archive_tradingload}
\title{Retrieve archived TRADINGLOAD data}
\usage{
nemwebR_archive_tradingload(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of TRADINGLOAD data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
TRADINGLOAD data contains historical 30-minute generation quantities for all scheduled and non-scheduled generators in the NEM.

Archive data is available from July 2009 to 30 September 2021 when 5MS was implemented. In order to retrieve newer data you will need to use the nemwebR_current_DISPATCHLOAD function.
}
\examples{
nemwebR_archive_tradingload(20210101)

}
