% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_tradingprice.R
\name{nemwebR_archive_tradingprice}
\alias{nemwebR_archive_tradingprice}
\title{Retrieve archived TRADINGPRICE data}
\usage{
nemwebR_archive_tradingprice(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of TRADINGPRICE data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
TRADINGPRICE data contains historical 30-minute prices for all regions in the NEM.

Archive data is available from July 2009 up to approximately one month ago. In order to retrieve newer data you will need to use the nemwebR_current_TRADINGPRICE function.
}
\examples{
nemwebR_archive_tradingprice(20210101)

}
