% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_archive_tradingregionsum.R
\name{nemwebR_archive_tradingregionsum}
\alias{nemwebR_archive_tradingregionsum}
\title{Retrieve archived TRADINGREGIONSUM data}
\usage{
nemwebR_archive_tradingregionsum(datestring)
}
\arguments{
\item{datestring}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns one month of TRADINGREGIONSUM data from AEMO's NEMWeb as specified by the datestring argument
}
\details{
TRADINGREGIONSUM data contains historical 30-minute demand and non-scheduled generation quantities for all NEM regions.

Archive data is available from July 2009 to 30 September 2021 when 5MS was implemented. In order to retrieve newer data you will need to use the nemwebR_current_DISPATCHREGIONSUM function.
}
\examples{
nemwebR_archive_tradingregionsum(20210101)

}
