% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemwebR_current_dispatchprice.R
\name{nemwebR_current_dispatchprice}
\alias{nemwebR_current_dispatchprice}
\title{Retrieve current DISPATCHPRICE data}
\usage{
nemwebR_current_dispatchprice(start_date, end_date)
}
\arguments{
\item{start_date}{integer of the form YYYYMMDD}

\item{end_date}{integer of the form YYYYMMDD}
}
\value{
A data frame
}
\description{
This function returns DISPATCHPRICE data from AEMO's NEMWeb as specified by the start and end date arguments.
}
\details{
DISPATCHPRICE data contains 5-minute regional pricing data for energy and FCAS markets.

Current data is available from yesterday to approximately two months ago. In order to retrieve older data you will need to use the \code{nemwebR_archive_dispatchprice()} function.
}
\examples{
nemwebR_current_dispatchprice(20210101, 20210131)


}
